feat: implement logical and physical planner
It is not complete yet but should follow a default way to handle
queries. The logical planner sorts out what a overall plan should be.
The physical plan will do more work to figure out if I should ready from
an index, etc. There is also an extra step that will use the
PlannerContext to keep track of table size and stats to determine how to
organize and restructure the physiocal query for maximum performance and
memory.